# Improve `Node#each_recursive` performance